refactor: create spec only crate without default features
With this change when you use keramik-operator as a dependency without
its default features you get a crate that only contains the spec types
with minimal deps.
