+++ Makefile
@@ -1,12 +1,16 @@
+CARGO = CARGO_REGISTRIES_CRATES_IO_PROTOCOL=sparse RUSTFLAGS="-D warnings" cargo
+
 .PHONY: all
 all: build check-fmt check-clippy test
 
 .PHONY: test
 test:
 	# Test with default features
-	cargo test --locked
+	${CARGO} test --locked
 	# Test with all features
-	cargo test --locked --all-features
+	${CARGO} test --locked --all-features
+	# Test without default features
+	${CARGO} test --locked --no-default-features
 
 .PHONY: check-fmt
 check-fmt:
@@ -15,17 +19,17 @@ check-fmt:
 .PHONY: check-clippy
 check-clippy:
 	# Check with default features
-	cargo clippy --workspace --all-targets -- -D warnings
+	${CARGO} clippy --workspace --all-targets
 	# Check with all features
-	cargo clippy --workspace --all-targets --all-features -- -D warnings
+	${CARGO} clippy --workspace --all-targets --all-features
 
 .PHONY: build
 build: runner operator
 
 .PHONY: runner
 runner:
-	CARGO_REGISTRIES_CRATES_IO_PROTOCOL=sparse RUSTFLAGS="-D warnings" cargo build --bin keramik-runner --release --locked  --config net.git-fetch-with-cli=true
+	${CARGO} build --bin keramik-runner --release --locked  --config net.git-fetch-with-cli=true
 
 .PHONY: operator
 operator:
-	CARGO_REGISTRIES_CRATES_IO_PROTOCOL=sparse RUSTFLAGS="-D warnings" cargo build --bin keramik-operator --release --locked  --config net.git-fetch-with-cli=true
+	${CARGO} build --bin keramik-operator --release --locked  --config net.git-fetch-with-cli=true
