+++ common/Cargo.toml
@@ -4,18 +4,23 @@ version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[features]
+default = []
+telemetry = [
+    "opentelemetry",
+    "opentelemetry-otlp",
+    "tracing-opentelemetry",
+    "tracing-subscriber",
+    "tracing",
+]
 
 [dependencies]
-actix-web = "4.3.1"
 anyhow.workspace = true
 gethostname = "0.4.2"
-opentelemetry-otlp.workspace = true
-opentelemetry.workspace = true
+opentelemetry-otlp = { workspace = true, optional = true }
+opentelemetry = { workspace = true, optional = true }
 schemars.workspace = true
 serde.workspace = true
-serde_json.workspace = true
-tokio.workspace = true
-tonic.workspace = true
-tracing-opentelemetry.workspace = true
-tracing-subscriber.workspace = true
-tracing.workspace = true
+tracing-opentelemetry = { workspace = true, optional = true }
+tracing-subscriber = { workspace = true, optional = true }
+tracing = { workspace = true, optional = true }
