+++ common/src/lib.rs
@@ -1,4 +1,5 @@
 //! Provides types and functions that common to both the runner and operator.
 #![deny(missing_docs)]
 pub mod peer_info;
+#[cfg(feature = "telemetry")]
 pub mod telemetry;
